Add helpers to approx equate the int32. (#8666)
* Add helpers to approx equate the int32.

In case of various platforms the ebc calculations can be one-off.
Which is not ideal, but in the grand scheme of things probably not as important as long as it's consistent (given the arbitrary choice of how we compute it anyway).
So add helpers to equate ±1.

* fixes

+++ pkg/autoscaler/scaling/autoscaler.go
@@ -262,10 +262,7 @@ func (a *autoscaler) Scale(ctx context.Context, now time.Time) ScaleResult {
 		excessBCF = 0
 		// numAct stays at MinActivators, only needed to scale from 0.
 	case a.deciderSpec.TargetBurstCapacity > 0:
-		// Extra float64 cast disables fused multiply-subtract to force identical behavior on
-		// all platforms. See floating point section in https://golang.org/ref/spec#Operators.
-		// nolint:unconvert
-		totCap := float64(float64(originalReadyPodsCount) * a.deciderSpec.TotalValue)
+		totCap := float64(originalReadyPodsCount) * a.deciderSpec.TotalValue
 		excessBCF = math.Floor(totCap - a.deciderSpec.TargetBurstCapacity - observedPanicValue)
 		numAct = int32(math.Max(MinActivators,
 			math.Ceil((totCap+a.deciderSpec.TargetBurstCapacity)/a.deciderSpec.ActivatorCapacity)))
